requirements.txt
openai
openai-agents
python_dotenv
markitdown  # parse any files to md
semantic-kernel
promptflow
https://github.com/Azure-Samples/aoai-realtime-audio-sdk/releases/download/py%2Fv0.5.3/rtclient-0.5.3.tar.gz
streamlit
audio-recorder-streamlit
azure-ai-inference
azure-ai-projects
azure-ai-agents
azure-identity
#azure-search-documents==11.5.2
azure-search-documents==11.7.0b1
azure-storage-blob
aiohttp
gradio
